Create Finished_Sample_Check module
The original module for this whole project (I'm sure there are structural issues with it). There are many ways for this particular problem to be solved, this is what I chose and worked on until finished. It's in VBA, it has classes, structures, modules, and a few ByRef or ByVal variables. It works and has proven useful.